% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis-stream-id.R
\name{as.character.redis_stream_id}
\alias{as.character.redis_stream_id}
\title{Redis Stream Identifier to Time Stamp and Sequence Number}
\usage{
\method{as.character}{redis_stream_id}(x)
}
\arguments{
\item{x}{Redis stream identifier.}
}
\value{
Vector of two character scalars: the Redis time stamp in
milliseconds, the zero-based sequence number.
}
\description{
Splits the stream identifier by its dash delimiter.
}
