% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis-stream-id.R
\name{as.POSIXct.redis_stream_id}
\alias{as.POSIXct.redis_stream_id}
\title{Redis Stream Identifier to POSIX Calendar Time}
\usage{
\method{as.POSIXct}{redis_stream_id}(x, ...)
}
\arguments{
\item{x}{Redis stream identifier.}
}
\value{
POSIX calendar time.
}
\description{
Redis Stream Identifier to POSIX Calendar Time
}
