% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis-stream-id.R
\name{as.redis_stream_id.character}
\alias{as.redis_stream_id.character}
\title{Character to Redis Stream Identifier}
\usage{
as.redis_stream_id.character(x)
}
\arguments{
\item{x}{Character vector of length 1.}
}
\value{
Redis stream identifier, the same character vector of length 1 but
with class \code{redis_stream_id}.
}
\description{
Character to Redis Stream Identifier
}
