% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalsha.R
\name{eval_sha}
\alias{eval_sha}
\title{Load Scripts by SHA for Evaluation}
\usage{
eval_sha(named_scripts, ...)
}
\arguments{
\item{named_scripts}{Named list of Lua scripts}

\item{...}{
  Arguments passed on to \code{\link[=script_load.evalsha]{script_load.evalsha}}
  \describe{
    \item{\code{redis}}{High-level Redis interface}
  }}
}
\value{
List of named wrapper functions, one per Lua script
}
\description{
Load Scripts by SHA for Evaluation
}
