% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Redis High-Level Interface}
\usage{
hi(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[redux:hiredis]{redux::hiredis}}
  \describe{
    \item{\code{version}}{Version of the interface to generate.  If given as
a string to numeric version, then only commands that exist up to
that version will be included.  If given as \code{TRUE}, then we
will query the Redis server (with \code{INFO}) and extract the
version number that way.}
  }}
}
\value{
Redis interface facade, R6 \code{redis_api} instance
}
\description{
Redis High-Level Interface
}
\examples{
\dontrun{
redis <- redisr::hi()
redis$PING()
#> [Redis: PONG]
}
}
