% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inst.R
\name{read.inst_lua}
\alias{read.inst_lua}
\title{Reads All Package-Installed Lua Scripts}
\usage{
read.inst_lua(package = "redisr", ...)
}
\arguments{
\item{...}{Script base-name filters; reads \emph{all} scripts if no filters}
}
\value{
Named list of character vectors after reading the package-installed
Lua scripts. The names exclude the package path and the \code{lua}
file-name extension.
}
\description{
Reads All Package-Installed Lua Scripts
}
