% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{scripts}
\alias{scripts}
\title{Redis High-Level Interface to Lua Scripts}
\usage{
scripts(x, ...)
}
\arguments{
\item{x}{Redis high-level interface}

\item{...}{Extra parameters, including:
\itemize{
\item \code{scripts} for named Lua scripts to load
}}
}
\description{
Loads a list of scripts by name and SHA ready for subsequent
evaluation.
}
