% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{strip.raw}
\alias{strip.raw}
\title{Strip Raw Prefix}
\usage{
strip.raw(x, pre = c(charToRaw("X\\n"), as.raw(0L)))
}
\arguments{
\item{x}{Raw vector to strip}

\item{pre}{Prefix to strip off if present, a raw vector}
}
\value{
Stripped raw vector
}
\description{
Useful for handling raw reply vectors.
}
\examples{
library(redisr)
strip.raw(c(charToRaw("X\n"), as.raw(0), charToRaw("abc")))
}
