% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{scripts}
\alias{scripts}
\title{Redis High-Level Interface Lua Scripts}
\usage{
scripts(x, ...)
}
\arguments{
\item{x}{Redis interface}

\item{...}{Extra parameters}
}
\description{
Redis High-Level Interface Lua Scripts
}
